dreammesh/requirements.txt
rich
tqdm
ninja
#torch
torch-utils
numpy 
scipy
click
lpips
pandas
trimesh
PyMCubes
torch-ema
dearpygui
packaging
matplotlib
tensorboardX
opencv-python
imageio
imageio-ffmpeg
# pyfqmr
pymeshlab
# torch_scatter
xatlas
scikit-learn
# git+https://github.com/NVlabs/nvdiffrast/
# pip install pytorch3d -f https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py38_cu102_pyt1121/download.html
# pip install --no-index --no-cache-dir pytorch3d -f https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py38_cu113_pyt1121/download.html
# pip install --no-index --no-cache-dir pytorch3d -f https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py39_cu113_pyt1121/download.html